+++ vite.config.ts
@@ -3,7 +3,7 @@ import dts from 'vite-plugin-dts';
 
 export default defineConfig({
   plugins: [
-    dts({ rollupTypes: true }),
+    dts({ insertTypesEntry: true }),
   ],
   build: {
     lib: {